bunch of maintenance stuff: * deleted old project preferences * enabled more warning checks * fixed warnings from new checks * got rid of labels, they're evil! :-P

+++ src/Action.h
@@ -64,7 +64,7 @@
 // NSMenu delegates
 - (BOOL)menu:(NSMenu *)menu updateItem:(NSMenuItem *)item atIndex:(int)index shouldCancel:(BOOL)shouldCancel;
 - (BOOL)menuHasKeyEquivalent:(NSMenu *)menu forEvent:(NSEvent *)event target:(id *)target action:(SEL *)action;
-- (int)numberOfItemsInMenu:(NSMenu *)menu;
+- (NSUInteger)numberOfItemsInMenu:(NSMenu *)menu;
 
 
 @end
